+++ src/main/java/kr/co/anabada/user/mapper/UserMapper.java
@@ -38,6 +38,7 @@ public interface UserMapper {
     //회원 탈퇴기능
     @Update("UPDATE alluser SET userStatus = #{userStatus}, userDeactiveDate = #{userDeactiveDate} WHERE userNo = #{userNo}")
     void updateUserStatus(@Param("userNo") int userNo, @Param("userStatus") String userStatus, @Param("userDeactiveDate") LocalDateTime userDeactiveDate);
+    
     @Select("SELECT * FORM alluser WHERE userNo = #{userNo}")
     User selectAllUserByUserNo(@Param("userNo") int userNo);
